update deployment guard and tests
