+++ kvm-compose/testbedos-server/src/state_evaluation/models.rs
@@ -4,7 +4,6 @@ use futures_util::future::join_all;
 use futures_util::{stream, StreamExt, TryStreamExt};
 use kvm_compose_lib::state::schema::StateNetwork;
 use kvm_compose_lib::state::state_evaluation::{EvaluateState, StateComponentStatus};
-use serde_json::{json, Value};
 use std::collections::HashMap;
 use std::sync::Arc;
